+++ brewtils/rest/system_client.py
@@ -360,7 +360,7 @@ class SystemClient(object):
                 _system_display=self._system.display_name,
                 _output_type=self._commands[command_name].output_type,
                 _instance_name=self._default_instance,
-                **kwargs
+                **kwargs,
             )
         else:
             raise AttributeError(
@@ -510,6 +510,25 @@ class SystemClient(object):
         if raise_on_error and request.status == "ERROR":
             raise RequestFailedError(request)
 
+        # Support cross-server parent/child requests. Add parent if request has different host.
+        if request.parent is None and (
+            brewtils.plugin.CONFIG.bg_host.upper()
+            != self._easy_client.client.bg_host.upper()
+            or brewtils.plugin.CONFIG.bg_port != self._easy_client.client.bg_port
+            or brewtils.plugin.CONFIG.bg_url_prefix
+            != self._easy_client.client.bg_url_prefix
+        ):
+            request.parent = getattr(
+                brewtils.plugin.request_context, "current_request", None
+            )
+            request.has_parent = True
+            ec = EasyClient(
+                bg_host=brewtils.plugin.CONFIG.bg_host,
+                bg_port=brewtils.plugin.CONFIG.bg_port,
+                bg_url_prefix=brewtils.plugin.CONFIG.bg_url_prefix,
+            )
+            ec.put_request(request)
+
         return request
 
     def _get_parent_for_request(self):
@@ -526,8 +545,7 @@ class SystemClient(object):
             self._logger.warning(
                 "A parent request was found, but the destination beer-garden "
                 "appears to be different than the beer-garden to which this plugin "
-                "is assigned. Cross-server parent/child requests are not supported "
-                "at this time. Removing the parent context so the request doesn't fail."
+                "is assigned. Removing the parent context so the request doesn't fail."
             )
             return None